==> 5. build_word2vec_model.ipynb <==
import pandas as pd
from pandas import DataFrame
import re
import nltk
import logging
from gensim.models import word2vec
logging.basicConfig(format='%(asctime)s : %(levelname)s : %(message)s', level=logging.INFO)

nyt_npr_gr = pd.read_csv('intermediary/nyt_npr_gr.csv')
desc = nyt_npr_gr['desc']

# clean descriptions and parse them into sentences made up of word lists
def sent_to_words(sent):
    sent = re.sub(r'[^a-z]', ' ', sent.lower())
    words = sent.split()
    return words

sent_tokenizer = nltk.data.load('tokenizers/punkt/english.pickle')
def desc_to_sents(desc):
    desc = desc.decode('utf-8')
    sents = sent_tokenizer.tokenize(desc.strip())
    sents = [sent_to_words(sent) for sent in sents]
    return sents

desc = [desc_to_sents(d) for d in desc]

desc = sum(desc, [])
len(desc)

# train a word vector model
word2vec_model = word2vec.Word2Vec(desc, workers = 4, min_count = 10, size = 300, window = 50, sample = .001, seed = 2015)

# test the model with a couple examples
word2vec_model.most_similar('harry')

word2vec_model.most_similar(positive = ['king', 'woman'], negative = ['man'])

# save the model
word2vec_model.init_sims(replace = True)
word2vec_model.save('intermediary/word2vec_model')

# combine the word vector matrix with the chosen vocabulary
word2vec_matrix = DataFrame(word2vec_model.syn0, index = word2vec_model.index2word)

# remove stop words from the matrix
stop_words = nltk.corpus.stopwords.words('english')
stop_words = set(stop_words) & set(word2vec_model.index2word)
print stop_words

word2vec_matrix = word2vec_matrix.drop(stop_words)
print '\n', word2vec_model.syn0.shape, word2vec_matrix.shape

word2vec_matrix.head()

# hand off to r
word2vec_matrix.to_csv('intermediary/word2vec_matrix.csv')
